Add child layout controls to the dimensions panel
